requirements.txt
streamlit 
opencv-python 
pillow 
transformers
opencv-python-headless 
langchain_groq
requests
python-dotenv
torch
yt-dlp
blip
ffmpeg
plotly
youtube_transcript_api